Répartir le code en plusieurs modules
